+++ src/number/utilities.ts
@@ -1,3 +1,5 @@
+import { isNumber } from '../guards/primitives';
+import { isNumericString } from '../guards/specials';
 import type { Numeric } from '../types/index';
 import { CURRENCY_LOCALES } from './constants';
 import type { CurrencyCode, LocaleCode } from './types';
@@ -93,4 +95,20 @@ export const getOrdinal = (num: Numeric, withNumber = true): string => {
 	}
 
 	return withNumber ? String(num).concat(suffix) : suffix;
+};
+
+/**
+ * * Normalize a number or numeric string to a number.
+ * @description
+ * This function checks if the input is a number or a numeric string and converts it to a number.
+ * If the input is not a valid number or numeric string, it returns undefined.
+ * @param num - The number to normalize.
+ * @returns The normalized number or undefined if the input is not a valid number or numeric string.
+ */
+export const normalizeNumber = (num: unknown) => {
+	return (
+		isNumber(num) ? num
+		: isNumericString(num) ? Number(num)
+		: undefined
+	);
 };